Implement quad tree generation (division_and_initialization)
This implements Algorithm 1 (DivisionAndInitialization) of the ES paper.

Note that due to Rust's strict aliasing I modified the algorithm to use
recursion, whereas the original algorithm from the paper uses an
iterated approach for DivisionAndInitialization. The recursion depth
should not be an issue as the number of nodes grows with 4^depth. This
means that for a depth of 10 we already have one millon nodes. I hardly
believe we ever want to have such a high resolution!